ZPR lab
V1 of the ZPR livelab. I still need to add video of doing the whole lab and then short videos of the main parts of the labs. Will get that to youtube and put in links here soon.

Also will need the commands to auto create instances and small ATP database with private endpoint.  These 3 items should be in the sam compartment and share a VCN to keep it simple.