A couple more images  (#74)
* ZPR lab

V1 of the ZPR livelab. I still need to add video of doing the whole lab and then short videos of the main parts of the labs. Will get that to youtube and put in links here soon.

Also will need the commands to auto create instances and small ATP database with private endpoint.  These 3 items should be in the sam compartment and share a VCN to keep it simple.

* clean up

Clean up text to meet lint needs and rename the files to usse dash not underlines.

* PT Review

* small updates

minor clean up. Removed the desktop folder as sandbox and tenancy only

* minor text update

* Updates

Based on feedback I have added more details in places, better security attribute names and values. Also blacked out any ip addresses etc

* minor changes

* correct lab 1 title for sandbox

title was wrong

* improving the text

* Minor updates

A couple more images and updated text to help be more clear to the users.

---------

Co-authored-by: Paul Toal <paul.toal@oracle.com>
Co-authored-by: Dan Wiliams <127415766+dannymgithub@users.noreply.github.com>